separate bottomUpTable from YYBottomSheet
separate BottomTableView from YYBottomSheet to add another bottomSheet component
refs #29
